% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exomePeak2_hg19.R
\name{exomePeak2_hg19}
\alias{exomePeak2_hg19}
\title{Produce code for Peak Calling with exomePeak2}
\usage{
exomePeak2_hg19(
  coldata,
  bam_dir,
  front_name,
  single_base = FALSE,
  GC_correct = TRUE,
  correct_GC_bg = FALSE,
  background = "Gaussian_mixture",
  qtnorm = TRUE
)
}
\description{
Produce code for Peak Calling with exomePeak2
}
