% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exomePeak2_PC_onestep.R
\name{exomePeak2_PC_onestep}
\alias{exomePeak2_PC_onestep}
\title{Produce code for Peak Calling with exomePeak2}
\usage{
exomePeak2_PC_onestep(coldata, bam_dir, front_name, single_base = FALSE,
  ...)
}
\description{
Produce code for Peak Calling with exomePeak2
}
