% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guitar_plot_Meth.R
\name{guitar_plot_Meth}
\alias{guitar_plot_Meth}
\title{Produce code for guitar plot after refined methylation call.}
\usage{
guitar_plot_Meth(pkc_rds_file, front_name, ...)
}
\description{
Produce code for guitar plot after refined methylation call.
}
