% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce_arg.R
\name{introduce_arg}
\alias{introduce_arg}
\title{Introduce argument into a code chunk.}
\usage{
introduce_arg(code, arguments)
}
\arguments{
\item{code}{a character vector containing the code.}

\item{arguments}{a character vector with names being the arguments and entries being the inputs.}
}
\description{
Introduce argument into a code chunk.
}
