% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_peak_calling.R
\name{parallel_peak_calling}
\alias{parallel_peak_calling}
\title{Run peak callings in parallel on bash sever}
\usage{
parallel_peak_calling(coldata, bam_dir = "./", save_dir = "./",
  autocreate = T, parallel_num = 9, pc_function = exomePeak2_hg19,
  front_name = "pkc_", single_base = F, GC_correct = T,
  correct_GC_bg = F, background = "Gaussian_mixture")
}
\arguments{
\item{coldata}{a \code{data.frame} which it columns containing the design information.}

\item{bam_dir}{the directory containing the bam files.}

\item{save_dir}{the directory to save the resulting RDS files of the \link{\code{summarizedExomePeaks}} object.}

\item{autocreate}{a logical, TRUE if create the \code{save_dir} when not finded.}

\item{parallel_num}{a numeric indicate the number of parallel process created; default = 6.}

\item{pc_function}{a peak calling function that takes the segragated coldata as input, and take a code string as output.}

\item{front_name}{a character string define the prefixes of the saved files.}

\item{...}{arguments that passed to the peak calling function.}
}
\value{
multiple \link{\code{summarizedExomePeaks}} objects or others organized by experimental design.
}
\description{
Run peak callings in parallel on bash sever
}
\details{
The coldata will be splitted only through the column of "Experiment".
}
\examples{
ColData_hg19 <- read.csv("coldata_hg19.csv")
parallel_peak_calling(
coldata = ColData_hg19[ColData_hg19$Perturbation == "C",],
bam_dir = "~/merip_bam",
save_dir = "~/exomepeak2_test",
parallel_num = 9,
pc_function = exomePeak2_PC
)
}
