% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_report.R
\name{parallel_report}
\alias{parallel_report}
\title{Run reports on peak callings in parallel with bash sever.}
\usage{
parallel_report(
  pkc_dir = "~/exomepeak2_test",
  pkc_prefix = "pkc_",
  pkc_suffix = ".rds",
  autocreate = TRUE,
  save_dir = "~/exomepeak2_GC_peaks",
  parallel_num = 9,
  report_function = reads_GC_plot,
  front_name = "GCreads_"
)
}
\arguments{
\item{pkc_dir}{the directory containing the Peak calling results in RDS format.}

\item{pkc_prefix}{a character of the prefix of the peakcalling result files; default "pkc_".}

\item{pkc_suffix}{a character of the suffix of the peakcalling result files; default ".rds".}

\item{autocreate}{a logical, TRUE if create the \code{save_dir} when it does not exist; default TRUE.}

\item{save_dir}{the directory to save the reported files.}

\item{parallel_num}{an integer indicates the number of parallel process created; default = 6.}

\item{report_function}{a report function that takes the peak calling result as the input, and take a code string as output.}

\item{front_name}{a character string define the prefixes of the saved files.}

\item{...}{arguments that can be passed to the report function.}
}
\value{
multiple report files on peak calling results, which will use the same middle name.
}
\description{
Run reports on peak callings in parallel with bash sever.
}
\examples{
parallel_report(
 pkc_dir = "~/exomepeak2_test"
 pkc_prefix = "pkc_",
 pkc_suffix = ".rds",
 save_dir = "~/exomepeak2_GC_peaks",
 parallel_num = 9,
 report_function = reads_GC_plot
)
}
