% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRC_df.R
\name{PRC_df}
\alias{PRC_df}
\title{Generate data.frame for the precision recall curve.}
\usage{
PRC_df(result_lst, gt_gr, N = 60, group_name = "prc")
}
\arguments{
\item{result_lst}{a list generated by the function \code{get_predictive_values}.}

\item{gt_gr}{a GRanges for the ground truce of positive methylation sites, recommended to be all single based in width.}

\item{N}{number of points returned for each curve, default = 60.}

\item{group_name}{a character for the name of the group returned, default = "prc".}
}
\description{
Generate data.frame for the precision recall curve.
}
\examples{
whistle_predict <- readRDS("predictedResults.rds")
whistle_gr <- readRDS("exon_DRACH.rds")
whistle_gr$prob <- whistle_predict
result_lst <- readRDS("hESC_C_noGC.rds")
gt_gr <- whistle_gr[whistle_gr$prob > 0.5]
plot_df <- PRC_df(result_lst = result_lst, gt_gr = gt_gr, N = 100, group_name = "no GC")

}
