% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guitar_plot.R, R/reads_GC_plot.R
\name{reads_GC_plot}
\alias{reads_GC_plot}
\alias{reads_GC_plot}
\title{Produce code for guitar plot}
\usage{
reads_GC_plot(pkc_rds_file, front_name, ...)

reads_GC_plot(pkc_rds_file, front_name, ...)
}
