% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnohup.R
\name{Rnohup}
\alias{Rnohup}
\title{submit a bash commands on server background}
\usage{
Rnohup(command, bashname = "X")
}
\arguments{
\item{command}{character vector of the R command used.}

\item{bashname}{the prefix name of files saved under bash.}
}
\description{
submit a bash commands on server background
}
