% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_macs2_code.R
\name{write_macs2_code}
\alias{write_macs2_code}
\title{Write code for macs2 (only for auprc purpose).}
\usage{
write_macs2_code(bam_ip, bam_input, name)
}
\arguments{
\item{bam_ip}{a vector containing the BAM files for IP samples.}

\item{bam_input}{a vector containing the BAM files for input samples.}

\item{name}{a character for the name of the output sh file and peak calling file.}
}
\value{
A file named by \code{name}.
}
\description{
Write code for macs2 (only for auprc purpose).
}
\examples{
OneStep_PRC(bam_ip = c("bam_local/SRR1035214_sorted.bam",
                       "bam_local/SRR1035216_sorted.bam",
                       "bam_local/SRR1035222_sorted.bam",
                       "bam_local/SRR1035224_sorted.bam"),
            bam_input = c("bam_local/SRR1035213_sorted.bam",
                          "bam_local/SRR1035215_sorted.bam",
                          "bam_local/SRR1035221_sorted.bam",
                          "bam_local/SRR1035223_sorted.bam"),
            name = "hNPC_control")

}
